Struct ariadne::Source

pub struct Source { /* private fields */ }

A type representing a single source that may be referred to by Spans.

In most cases, a source is a single input file.

Implementations

impl Source

pub fn len(&self) -> usize

Get the length of the total number of characters in the source.

pub fn chars(&self) -> impl Iterator<Item = char> + '_

Return an iterator over the characters in the source.

pub fn line(&self, idx: usize) -> Option<&Line>

Get access to a specific, zero-indexed Line.

pub fn lines(&self) -> impl ExactSizeIterator<Item = &Line> + '_

Return an iterator over the Lines in this source.

pub fn get_offset_line(&self, offset: usize) -> Option<(&Line, usize, usize)>

Get the line that the given offset appears on, and the line/column numbers of the offset.

Note that the line/column numbers are zero-indexed.

pub fn get_line_range<S: Span>(&self, span: &S) -> Range<usize>

Get the range of lines that this span runs across.

The resulting range is guaranteed to contain valid line indices (i.e: those that can be used for Source::line).