Trait ariadne::Fmt

pub trait Fmt: Sized {
    fn fg<C: Into<Option<Color>>>(self, color: C) -> Foreground<Self> { ... }
    fn bg<C: Into<Option<Color>>>(self, color: C) -> Background<Self> { ... }
}

A trait used to add formatting attributes to displayable items.

Attributes specified through this trait are not composable (i.e: the behaviour of two nested attributes each with a conflicting attribute is left unspecified).

Provided Methods

fn fg<C: Into<Option<Color>>>(self, color: C) -> Foreground<Self>

Give this value the specified foreground colour

fn bg<C: Into<Option<Color>>>(self, color: C) -> Background<Self>

Give this value the specified background colour